# backend/swagger/swagger.yaml
openapi: 3.1.0
info:
  title: ASAP HVAC Sample API
  version: 0.2.0
  description: |
    Sample REST API for the ASAP HVAC web demo. Provides CRUD support for
    HVAC operational data with pagination, filtering, and validation via Zod
    schemas. Designed to be upgraded to relational storage in future iterations.
servers:
  - url: http://localhost:4000
    description: Local development server
paths:
...
  /api/maintenance-plans:
    get:
      summary: List maintenance plans
      tags: [Maintenance plans]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: planTier
          in: query
          description: Filter by plan tier label.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of maintenance plans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMaintenancePlan'
    post:
      summary: Create a maintenance plan
      tags: [Maintenance plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenancePlanInput'
      responses:
        '201':
          description: Newly created maintenance plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenancePlan'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/maintenance-plans/{id}:
    parameters:
      - $ref: '#/components/parameters/MaintenancePlanId'
    get:
      summary: Retrieve a maintenance plan
      tags: [Maintenance plans]
      responses:
        '200':
          description: Maintenance plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenancePlan'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update maintenance plan
      tags: [Maintenance plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenancePlanUpdate'
      responses:
        '200':
          description: Updated maintenance plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenancePlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a maintenance plan
      tags: [Maintenance plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenancePlanUpdate'
      responses:
        '200':
          description: Updated maintenance plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenancePlan'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a maintenance plan
      tags: [Maintenance plans]
      responses:
        '204':
          description: Maintenance plan deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/invoices:
    get:
      summary: List invoices
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: finalQuoteId
          in: query
          description: Filter by final quote identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: paid
          in: query
          description: Filter by invoice payment status.
          schema:
            type: array
            items:
              type: string
              enum: ['true', 'false']
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of invoices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInvoice'
    post:
      summary: Create an invoice
      tags: [Invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '201':
          description: Newly created invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/invoices/{id}:
    parameters:
      - $ref: '#/components/parameters/InvoiceId'
    get:
      summary: Retrieve an invoice
      tags: [Invoices]
      responses:
        '200':
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update an invoice
      tags: [Invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update an invoice
      tags: [Invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdate'
      responses:
        '200':
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete an invoice
      tags: [Invoices]
      responses:
        '204':
          description: Invoice deleted successfully
        '400':
          description: Invoice cannot be deleted while related payments exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/payments:
    get:
      summary: List payments
      tags: [Payments]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: invoiceId
          in: query
          description: Filter by invoice identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: method
          in: query
          description: Filter by payment method label.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPayment'
    post:
      summary: Create a payment
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
      responses:
        '201':
          description: Newly created payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/payments/{id}:
    parameters:
      - $ref: '#/components/parameters/PaymentId'
    get:
      summary: Retrieve a payment
      tags: [Payments]
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a payment
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdate'
      responses:
        '200':
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a payment
      tags: [Payments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentUpdate'
      responses:
        '200':
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a payment
      tags: [Payments]
      responses:
        '204':
          description: Payment deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /api/parts-orders:
    get:
      summary: List parts orders
      tags: [Parts orders]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: finalQuoteId
          in: query
          description: Filter by final quote identifier.
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: status
          in: query
          description: Filter by parts order status.
          schema:
            type: array
            items:
              type: string
              enum: [ordered, backordered, fulfilled, canceled]
          style: form
          explode: true
      responses:
        '200':
          description: Paginated list of parts orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPartsOrder'
    post:
      summary: Create a parts order
      tags: [Parts orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartsOrderInput'
      responses:
        '201':
          description: Newly created parts order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartsOrder'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/parts-orders/{id}:
    parameters:
      - $ref: '#/components/parameters/PartsOrderId'
    get:
      summary: Retrieve a parts order
      tags: [Parts orders]
      responses:
        '200':
          description: Parts order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartsOrder'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a parts order
      tags: [Parts orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartsOrderUpdate'
      responses:
        '200':
          description: Updated parts order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartsOrder'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Partially update a parts order
      tags: [Parts orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartsOrderUpdate'
      responses:
        '200':
          description: Updated parts order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartsOrder'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a parts order
      tags: [Parts orders]
      responses:
        '204':
          description: Parts order deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based indexing).
      schema:
        type: integer
        minimum: 1
    LimitParam:
      name: limit
      in: query
      description: Number of records per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: Sort expression in the format `<field>[:asc|desc]`.
      schema:
        type: string
...
    MaintenancePlanId:
      name: id
      in: path
      required: true
      description: Unique identifier for the maintenance plan
      schema:
        type: string
...
    InvoiceId:
      name: id
      in: path
      required: true
      description: Unique identifier for the invoice
      schema:
        type: string
    PaymentId:
      name: id
      in: path
      required: true
      description: Unique identifier for the payment
      schema:
        type: string
    PartsOrderId:
      name: id
      in: path
      required: true
      description: Unique identifier for the parts order
      schema:
        type: string
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: Customer email must be unique.
        details:
          description: Optional error metadata
          nullable: true
    ListMeta:
      type: object
      required: [total, page, limit, hasNextPage]
      properties:
        total:
          type: integer
          example: 45
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 25
        hasNextPage:
          type: boolean
          example: true
...
    PaginatedMaintenancePlan:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MaintenancePlan'
        meta:
          $ref: '#/components/schemas/ListMeta'
    MaintenancePlan:
      type: object
      required: [id, planTier, annualFee, includedServices, partsDiscountPct, extras]
      properties:
        id:
          type: string
        planTier:
          type: string
        annualFee:
          type: number
        includedServices:
          type: array
          items:
            type: string
        partsDiscountPct:
          type: integer
          minimum: 0
          maximum: 100
        extras:
          type: array
          items:
            type: string
    MaintenancePlanInput:
      type: object
      required: [planTier, annualFee, includedServices, partsDiscountPct]
      properties:
        id:
          type: string
        planTier:
          type: string
        annualFee:
          type: number
        includedServices:
          type: array
          items:
            type: string
        partsDiscountPct:
          type: integer
        extras:
          type: array
          items:
            type: string
    MaintenancePlanUpdate:
      allOf:
        - $ref: '#/components/schemas/MaintenancePlanInput'
...
    PaginatedInvoice:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Invoice'
        meta:
          $ref: '#/components/schemas/ListMeta'
    Invoice:
      type: object
      required: [id, finalQuoteId, amountDue, createdOn, dueDate, paid]
      properties:
        id:
          type: string
        finalQuoteId:
          type: string
        amountDue:
          type: number
        createdOn:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        paid:
          type: boolean
        paymentRef:
          type: string
          nullable: true
    InvoiceInput:
      type: object
      required: [finalQuoteId, amountDue, createdOn, dueDate, paid]
      properties:
        id:
          type: string
        finalQuoteId:
          type: string
        amountDue:
          type: number
        createdOn:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        paid:
          type: boolean
        paymentRef:
          type: string
          nullable: true
    InvoiceUpdate:
      allOf:
        - $ref: '#/components/schemas/InvoiceInput'
    PaginatedPayment:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        meta:
          $ref: '#/components/schemas/ListMeta'
    Payment:
      type: object
      required: [id, invoiceId, amount, paidOn, method]
      properties:
        id:
          type: string
        invoiceId:
          type: string
        amount:
          type: number
        paidOn:
          type: string
          format: date
        method:
          type: string
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    PaymentInput:
      type: object
      required: [invoiceId, amount, paidOn, method]
      properties:
        id:
          type: string
        invoiceId:
          type: string
        amount:
          type: number
        paidOn:
          type: string
          format: date
        method:
          type: string
        reference:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    PaymentUpdate:
      allOf:
        - $ref: '#/components/schemas/PaymentInput'
    PaginatedPartsOrder:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartsOrder'
        meta:
          $ref: '#/components/schemas/ListMeta'
    PartsOrder:
      type: object
      required: [id, finalQuoteId, items, totalCost, status]
      properties:
        id:
          type: string
        finalQuoteId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartsOrderItem'
        totalCost:
          type: number
        status:
          type: string
          enum: [ordered, backordered, fulfilled, canceled]
        etaDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    PartsOrderItem:
      type: object
      required: [partId, description, qty, costEach]
      properties:
        partId:
          type: string
        description:
          type: string
        qty:
          type: integer
        costEach:
          type: number
    PartsOrderInput:
      type: object
      required: [finalQuoteId, items, totalCost, status]
      properties:
        id:
          type: string
        finalQuoteId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartsOrderItem'
        totalCost:
          type: number
        status:
          type: string
          enum: [ordered, backordered, fulfilled, canceled]
        etaDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    PartsOrderUpdate:
      allOf:
        - $ref: '#/components/schemas/PartsOrderInput'